+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file. This projec
 - `changes associate` breaking when using an invalid revision [#107](https://github.com/Brickstertwo/git-commands/issues/107)
 - `changes unassociate --dry-run` printing a message during a no-op [#108](https://github.com/Brickstertwo/git-commands/issues/108)
 - `changes associate --upstream` failing with local upstreams [#112](https://github.com/Brickstertwo/git-commands/issues/112)
+- `settings list` failing with empty config files [#121](https://github.com/Brickstertwo/git-commands/issues/121)
 
 ## [v0.6.1][] - 2017-01-17
 ### Fixed